requirements.txt
pdfplumber
python-docx
spacy
pydantic
scikit-learn
streamlit
pytest
